requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.4
bcrypt==4.3.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# PostgreSQL 관련 패키지
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# JWT 토큰 처리
python-jose[cryptography]==3.3.0
